Migrate scrolling tabs to actual SemUI menu
